Move test files to proper directory
