+++ io/src/main/java/org/n52/io/handler/DatasetFactoryException.java
@@ -26,7 +26,7 @@
  * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
  * for more details.
  */
-package org.n52.io;
+package org.n52.io.handler;
 
 public class DatasetFactoryException extends Exception {
