+++ rest/src/main/java/org/n52/web/ctrl/TimeseriesDataController.java
@@ -46,9 +46,9 @@ import org.joda.time.Duration;
 import org.joda.time.Interval;
 import org.joda.time.Period;
 import org.n52.io.Constants;
-import org.n52.io.DatasetFactoryException;
 import org.n52.io.IntervalWithTimeZone;
 import org.n52.io.PreRenderingJob;
+import org.n52.io.handler.DatasetFactoryException;
 import org.n52.io.handler.DefaultIoFactory;
 import org.n52.io.handler.IoHandlerException;
 import org.n52.io.handler.IoHandlerFactory;
