+++ rest/src/main/java/org/n52/web/ctrl/data/DataController.java
@@ -45,8 +45,8 @@ import org.joda.time.Duration;
 import org.joda.time.Interval;
 import org.joda.time.Period;
 import org.n52.io.Constants;
-import org.n52.io.DatasetFactoryException;
 import org.n52.io.IntervalWithTimeZone;
+import org.n52.io.handler.DatasetFactoryException;
 import org.n52.io.handler.DefaultIoFactory;
 import org.n52.io.handler.IoHandlerFactory;
 import org.n52.io.handler.IoProcessChain;
